Жёсткий рефакторинг

+++ DoiFApp/Services/Workload/ExcelWorkloadWriter.cs
@@ -1,53 +1,41 @@
-﻿using DoiFApp.Data.Models;
-using DoiFApp.Data.Repo;
-using OfficeOpenXml;
+﻿using OfficeOpenXml;
 using System.IO;
 using OfficeOpenXml.Style;
 using System.Drawing;
 using DoiFApp.Utils;
 using System.Text;
+using DoiFApp.Services.Data;
 
-namespace DoiFApp.Services.Excel
+namespace DoiFApp.Services.Workload
 {
-    public class ExcelWorkSchedule(IRepo<LessonModel> lessonRepo) : IWorkSchedule
+    public class ExcelWorkloadWriter : IDataWriter<WorkloadData>
     {
-        private readonly IRepo<LessonModel> lessonRepo = lessonRepo;
-
-        private readonly Color saturdayColor = Color.FromArgb(255, 255, 192, 203); // low red
-        private readonly Color sundayColor = Color.FromArgb(255, 255, 0, 0); // red
-
-        private readonly Color koptevoColor = Color.FromArgb(255, 199, 199, 199); // grey
-        private readonly Color volginoColor = Color.FromArgb(255, 255, 255, 230); // yellow
-        private readonly Color otherColor = Color.FromArgb(255, 255, 230, 200); // orange
-        private readonly Color transitionColor = Color.FromArgb(255, 100, 245, 100); // green
-        private readonly Color withoutColor = Color.FromArgb(255, 0, 200, 200); // blue
-
-        public async Task Write(string path)
+        public Task<bool> Write(WorkloadData data, string path)
         {
+            if (!data.IsHolistic) return Task.FromResult(false);
+
             if (File.Exists(path))
                 File.Delete(path);
 
             using var package = new ExcelPackage(path);
             var worksheet = package.Workbook.Worksheets.Add("График");
 
-            var data = await lessonRepo.GetAll();
+            var teachersUnique = DataUtil.GetTeachers(data.Lessons!);
 
-            var teachersUnique = DataUtil.GetTeachers(data);
-
-            (int start, int end) teacherLine = (3, teachersUnique.Count + 2);
+            (int start, int end) = (3, teachersUnique.Count + 2);
 
             // header
             for (var i = 0; i < teachersUnique.Count; i++)
             {
                 var teacher = teachersUnique[i];
-                var tcell = worksheet.Cells[1, teacherLine.start + i];
+                var tcell = worksheet.Cells[1, start + i];
                 tcell.Value = teacher;
                 tcell.Style.Border.BorderAround(ExcelBorderStyle.Hair);
                 tcell.Style.TextRotation = 180;
             }
 
-            var startDayNum = data.First().Date.DayNumber;
-            var endDayNum = data.Last().Date.DayNumber;
+            var startDayNum = data.Lessons!.First().Date.DayNumber;
+            var endDayNum = data.Lessons!.Last().Date.DayNumber;
 
             for (var i = startDayNum; i < endDayNum; i++)
             {
@@ -59,12 +47,12 @@ namespace DoiFApp.Services.Excel
                 worksheet.Cells[tableVerticalIndex, 2].Value = curDay;
 
                 if (date.DayOfWeek == DayOfWeek.Saturday)
-                    DrawLine(worksheet, tableVerticalIndex, teacherLine.end, saturdayColor);
+                    DrawLine(worksheet, tableVerticalIndex, end, WorkloadHelper.SaturdayColor);
 
                 if (date.DayOfWeek == DayOfWeek.Sunday)
-                    DrawLine(worksheet, tableVerticalIndex, teacherLine.end, sundayColor);
+                    DrawLine(worksheet, tableVerticalIndex, end, WorkloadHelper.SundayColor);
 
-                var lessons = data.Where(l => l.Date == date);
+                var lessons = data.Lessons!.Where(l => l.Date == date);
 
                 if (lessons == null)
                     continue;
@@ -76,20 +64,20 @@ namespace DoiFApp.Services.Excel
                         var teacher = teachersUnique[j];
                         if (lesson.Teachers.Contains(teacher))
                         {
-                            var lessionsCell = worksheet.Cells[tableVerticalIndex, teacherLine.start + j];
+                            var lessionsCell = worksheet.Cells[tableVerticalIndex, start + j];
                             lessionsCell.Value += SwitchClassId(lesson.Time) + " ";
 
                             lessionsCell.Style.Fill.PatternType = ExcelFillStyle.Solid;
-                            lessionsCell.Style.Fill.BackgroundColor.SetColor(SelectCellColor(lesson.Auditoriums));
+                            lessionsCell.Style.Fill.BackgroundColor.SetColor(WorkloadHelper.SelectCellColor(lesson.Auditoriums));
                         }
                     }
                 }
             }
 
-            AddNotes(teacherLine.end + 1, worksheet);
+            AddNotes(end + 1, worksheet);
 
             // view
-            DoSquare(worksheet, 1, 1, endDayNum - startDayNum + 1, teacherLine.end, (range) =>
+            DoSquare(worksheet, 1, 1, endDayNum - startDayNum + 1, end, (range) =>
             {
                 range.Style.VerticalAlignment = ExcelVerticalAlignment.Center;
                 range.Style.HorizontalAlignment = ExcelHorizontalAlignment.Center;
@@ -108,17 +96,19 @@ namespace DoiFApp.Services.Excel
             });
 
             package.Save();
+
+            return Task.FromResult(true);
         }
 
-        private void AddNotes(int startNoteX, ExcelWorksheet worksheet)
+        private static void AddNotes(int startNoteX, ExcelWorksheet worksheet)
         {
             int startNoteY = 2, i = 0;
 
-            AddNote(worksheet, startNoteY + i++, startNoteX + 1, koptevoColor, "Коптево");
-            AddNote(worksheet, startNoteY + i++, startNoteX + 1, volginoColor, "Волгино");
-            AddNote(worksheet, startNoteY + i++, startNoteX + 1, otherColor, "Др. площадки");
-            AddNote(worksheet, startNoteY + i++, startNoteX + 1, withoutColor, "Без аудитории");
-            AddNote(worksheet, startNoteY + i++, startNoteX + 1, transitionColor, "Переезд");
+            AddNote(worksheet, startNoteY + i++, startNoteX + 1, WorkloadHelper.KoptevoColor, "Коптево");
+            AddNote(worksheet, startNoteY + i++, startNoteX + 1, WorkloadHelper.VolginoColor, "Волгино");
+            AddNote(worksheet, startNoteY + i++, startNoteX + 1, WorkloadHelper.OtherColor, "Др. площадки");
+            AddNote(worksheet, startNoteY + i++, startNoteX + 1, WorkloadHelper.WithoutColor, "Без аудитории");
+            AddNote(worksheet, startNoteY + i++, startNoteX + 1, WorkloadHelper.TransitionColor, "Переезд");
 
             i++;
 
@@ -134,21 +124,20 @@ namespace DoiFApp.Services.Excel
             AddNote(worksheet, startNoteY + i++, startNoteX + 1, "7", "20:00 - 21:30");
         }
 
-        private void AddNote(ExcelWorksheet worksheet, int y, int x, object key, object value)
+        private static void AddNote(ExcelWorksheet worksheet, int y, int x, object key, object value)
         {
             worksheet.Cells[y, x].Value = key;
             worksheet.Cells[y, x + 1].Value = value;
         }
 
-        private void AddNote(ExcelWorksheet worksheet, int y, int x, Color key, object value)
+        private static void AddNote(ExcelWorksheet worksheet, int y, int x, Color key, object value)
         {
             worksheet.Cells[y, x].Style.Fill.PatternType = ExcelFillStyle.Solid;
             worksheet.Cells[y, x].Style.Fill.BackgroundColor.SetColor(key);
             worksheet.Cells[y, x + 1].Value = value;
-
         }
 
-        private string SwitchClassId(string time)
+        private static string SwitchClassId(string time)
             => time switch
             {
                 "09:00 - 10:30" => "1",
@@ -163,7 +152,7 @@ namespace DoiFApp.Services.Excel
                 _ => "n"
             };
 
-        private void DrawLine(ExcelWorksheet worksheet, int vIndex, int hIndex, Color color)
+        private static void DrawLine(ExcelWorksheet worksheet, int vIndex, int hIndex, Color color)
         {
             var rangeString = $"A{vIndex}:{GetEndSymbol(hIndex)}{vIndex}";
             var range = worksheet.Cells[rangeString];
@@ -172,39 +161,15 @@ namespace DoiFApp.Services.Excel
             range.Style.Fill.BackgroundColor.SetColor(color);
         }
 
-        private void DoSquare(ExcelWorksheet worksheet, int vIndex1, int hIndex1, int vIndex2, int hIndex2, Action<ExcelRange> action)
+        private static void DoSquare(ExcelWorksheet worksheet, int vIndex1, int hIndex1, int vIndex2, int hIndex2, Action<ExcelRange> action)
             => action(worksheet.Cells[$"{GetEndSymbol(hIndex1)}{vIndex1}:{GetEndSymbol(hIndex2)}{vIndex2}"]);
 
-        private string GetEndSymbol(int hIndex)
+        private static string GetEndSymbol(int hIndex)
         {
             var startByte = Encoding.ASCII.GetBytes("A");
             startByte[0] += Convert.ToByte(hIndex - 1);
             return Encoding.ASCII.GetString(startByte);
         }
-
-        private Color SelectCellColor(List<string> auditoriums)
-        {
-            var auds = auditoriums.Where(a => a.Contains("к/"));
-
-            if (!auds.Any())
-                return withoutColor;
-
-            var suites = auds.Select(a => a.Split("к/")[0]).Distinct();
-            if (suites.Count() > 1)
-                return transitionColor;
-
-            return SwitchColorByAuditorium(suites.First());
-        }
-
-        private Color SwitchColorByAuditorium(string id)
-            => id switch
-            {
-                "1" => volginoColor,
-                "2" => volginoColor,
-                "3" => volginoColor,
-                "5" => koptevoColor,
-                _ => otherColor,
-            };
     }
 }
